fix(test): complete collision rename consumers
