+++ src/utils/timer-delay.ts
@@ -3,7 +3,6 @@ import { resolveSafeTimeoutDelayMs } from "../../packages/gateway-client/src/tim
 
 export {
   addSafeTimeoutDelayGraceMs,
-  MAX_SAFE_TIMEOUT_DELAY_MS,
   resolveSafeTimeoutDelayMs,
 } from "../../packages/gateway-client/src/timeouts.js";
